# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API",
        "description": "API documentation for the Ecommerce platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/orders/history": {
            "get": {
                "tags": [
                    "Order History"
                ],
                "summary": "Get all order history for the authenticated user",
                "operationId": "1511779fbf2fbc2c57a41daf55e7dd29",
                "responses": {
                    "200": {
                        "description": "Order history retrieved successfully"
                    },
                    "500": {
                        "description": "Order history retrieval error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/history/{id}": {
            "get": {
                "tags": [
                    "Order History"
                ],
                "summary": "Get specific order details",
                "operationId": "06885ac615091802f7b9423030e83a7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details retrieved successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "500": {
                        "description": "Order details retrieval error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/history/{id}/status": {
            "put": {
                "tags": [
                    "Order History"
                ],
                "summary": "Update the status of a specific order",
                "operationId": "e37f78dc96e1f6562004f759b26e001d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "shipped"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order status updated successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "500": {
                        "description": "Order status update error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/payments": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Process payment for an order",
                "operationId": "6ed7838e6c677c90f52dd3ba24f13269",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "pm_card_visa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment processed successfully"
                    },
                    "404": {
                        "description": "Order not found or already paid"
                    },
                    "500": {
                        "description": "Payment processing error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment details for a specific order",
                "operationId": "d5c266c64ee63f926a92a8d127a57e7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment details retrieved successfully"
                    },
                    "404": {
                        "description": "Payment not found"
                    },
                    "500": {
                        "description": "Payment retrieval error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}